+++ drivers/hmis/app/graphql/mutations/create_direct_upload.rb
@@ -7,7 +7,7 @@ module Mutations
     type Types::Uploads::DirectUploadType
 
     def resolve(input:)
-      blob = ActiveStorage::Blob.create_before_direct_upload!(input.to_h)
+      blob = ActiveStorage::Blob.create_before_direct_upload!(**input.to_h)
       {
         url: blob.service_url_for_direct_upload,
         # NOTE: we pass headers as JSON since they have no schema
